Moving colors function to the corresponding file
