+++ verticapy/learn/model_selection.py
@@ -36,7 +36,7 @@ from verticapy import vDataFrame
 from verticapy.utilities import *
 from verticapy.toolbox import *
 from verticapy.errors import *
-from verticapy.plotting._matplotlib import gen_colors
+from verticapy.plotting._colors import gen_colors
 from verticapy.learn.tools import does_model_exist
 from verticapy.learn.mlplot import plot_bubble_ml, plot_stepwise_ml, plot_importance
