+++ verticapy/plotting/_matplotlib/animated.py
@@ -29,7 +29,7 @@ import numpy as np
 from verticapy.utilities import *
 from verticapy.toolbox import executeSQL, quote_ident
 from verticapy.errors import ParameterError
-from verticapy.plotting._matplotlib.core import gen_cmap, gen_colors
+from verticapy.plotting._colors import gen_cmap, gen_colors
 
 # IPython - Optional
 try:
