+++ verticapy/plotting/_matplotlib/boxplot.py
@@ -24,7 +24,7 @@ import matplotlib.pyplot as plt
 from verticapy.utilities import *
 from verticapy.toolbox import executeSQL, quote_ident
 from verticapy.errors import ParameterError
-from verticapy.plotting._matplotlib.core import gen_colors
+from verticapy.plotting._colors import gen_colors
 
 
 def boxplot(
